fixed_webhook_bot.py
#!/usr/bin/env python3
"""
FIXED WEBHOOK BOT - All commands working
"""

import os
import logging
import asyncio
from flask import Flask, request
from telegram import Update
from telegram.ext import Application, CommandHandler, ContextTypes

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Configuration
BOT_TOKEN = "8415230764:AAF0Aaqb21Vkq9eWifB_wHDtkm37WrjJRcs"
WEBHOOK_URL = "https://bigbullscalptradingbot.herokuapp.com"

# Flask app
app = Flask(__name__)

# Create application
application = Application.builder().token(BOT_TOKEN).build()

async def start_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Start command handler"""
    logger.info(f"Start command from user {update.effective_user.id}")
    
    message = """
🎉 **BigBullScalpBot is LIVE!** 🎉

✅ F&O Trading Signals
✅ Live Market Analysis  
✅ Payment Integration
✅ 24/7 Automated Signals

**Monthly Subscription: ₹2,400**

🔧 **Test Commands:**
/test - Generate test signal
/about - About bot
/status - Bot status

**UPI Payment:**
📱 UPI ID: 81285083843@YBL
💰 Amount: ₹2,400

Your trading bot is now LIVE! 🚀
    """
    
    await update.message.reply_text(message, parse_mode='Markdown')

async def test_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Test signal command"""
    logger.info(f"Test command from user {update.effective_user.id}")
    
    message = """
🚨 **TEST SIGNAL GENERATED** 🚨

📈 **Symbol:** NSE:NIFTY  
🔥 **Type:** BUY
💰 **Entry:** ₹24,850.00
🎯 **Target:** ₹25,100.00  
🛑 **Stop Loss:** ₹24,600.00

⏰ **Generated:** Just now
🤖 **Status:** Test signal (System working!)

✅ **Signal generation system is operational!**

Try /about for more information.
    """
    
    await update.message.reply_text(message, parse_mode='Markdown')

async def about_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """About command"""
    logger.info(f"About command from user {update.effective_user.id}")
    
    message = """
🤖 **BigBullScalpBot**

🎯 **Features:**
• Live F&O trading signals
• Real-time market analysis  
• Risk management included
• 24/7 automated operation

💰 **Pricing:** ₹2,400/month
📱 **UPI ID:** 81285083843@YBL

📊 **Signal Types:**
• NIFTY F&O
• BANKNIFTY F&O  
• Blue chip stocks

**Status:** ✅ ONLINE & WORKING

⚠️ **Disclaimer:** Trading involves risk
    """
    
    await update.message.reply_text(message, parse_mode='Markdown')

async def status_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Status command"""
    logger.info(f"Status command from user {update.effective_user.id}")
    
    from datetime import datetime
    message = f"""
📊 **Bot Status Report**

🤖 **Bot:** ✅ Online & Responding
🌐 **Server:** ✅ Heroku Active  
🔗 **Webhook:** ✅ Configured
📡 **Telegram API:** ✅ Connected

**Commands Working:**
✅ /start - Main menu
✅ /test - Test signal
✅ /about - Bot info  
✅ /status - This status

**Last Check:** {datetime.now().strftime('%H:%M:%S')}

🎯 **All systems operational!**
    """
    
    await update.message.reply_text(message, parse_mode='Markdown')

# Add all handlers
application.add_handler(CommandHandler("start", start_command))
application.add_handler(CommandHandler("test", test_command))
application.add_handler(CommandHandler("about", about_command))
application.add_handler(CommandHandler("status", status_command))

@app.route('/')
def home():
    return """
    <h1>🤖 BigBullScalpBot Status</h1>
    <h2>✅ ONLINE & WORKING</h2>
    <p><strong>Bot Username:</strong> @Bigbulscalp_bot</p>
    <p><strong>Deployment:</strong> Webhook Active</p>
    <p><strong>Commands:</strong> /start, /test, /about, /status</p>
    <hr>
    <h3>🔧 Admin Links:</h3>
    <p><a href="/setwebhook">Set Webhook</a></p>
    <p><a href="/health">Health Check</a></p>
    """

@app.route('/webhook', methods=['POST'])
def webhook():
    """Handle webhook updates from Telegram"""
    try:
        # Get update data
        json_data = request.get_json(force=True)
        
        # Create update object
        update = Update.de_json(json_data, application.bot)
        
        # Process update with proper async handling
        loop = asyncio.new_event_loop()
        asyncio.set_event_loop(loop)
        
        try:
            loop.run_until_complete(application.process_update(update))
        finally:
            loop.close()
        
        logger.info(f"Processed update: {update.update_id}")
        return "OK"
        
    except Exception as e:
        logger.error(f"Webhook error: {e}")
        return f"ERROR: {e}", 500

@app.route('/health')
def health():
    return """
    <h2>🟢 Bot Health Check</h2>
    <p>Status: <strong>HEALTHY</strong></p>
    <p>All systems operational!</p>
    """, 200

@app.route('/setwebhook')
def set_webhook():
    """Set webhook URL - run this once after deployment"""
    try:
        webhook_url = f"{WEBHOOK_URL}/webhook"
        
        # Set webhook using requests (more reliable)
        import requests
        url = f"https://api.telegram.org/bot{BOT_TOKEN}/setWebhook"
        data = {'url': webhook_url}
        
        response = requests.post(url, data=data)
        result = response.json()
        
        if result.get('ok'):
            return f"""
            <h2>✅ Webhook Set Successfully!</h2>
            <p><strong>Webhook URL:</strong> {webhook_url}</p>
            <p><strong>Status:</strong> Active</p>
            <p>Your bot should now respond to all commands!</p>
            <hr>
            <p><a href="/">← Back to Status</a></p>
            """
        else:
            return f"""
            <h2>❌ Webhook Setup Failed</h2>
            <p><strong>Error:</strong> {result.get('description', 'Unknown error')}</p>
            <p>Please try again or check the logs.</p>
            """
            
    except Exception as e:
        return f"""
        <h2>❌ Error Setting Webhook</h2>
        <p><strong>Error:</strong> {e}</p>
        <p>Please check the configuration and try again.</p>
        """

if __name__ == '__main__':
    logger.info("🚀 Starting BigBullScalpBot with fixed webhook...")
    
    # Start Flask app
    port = int(os.environ.get('PORT', 5000))
    app.run(host='0.0.0.0', port=port, debug=False)