requirements.txt
python-telegram-bot==20.1
flask==2.3.3
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0